fix model Settings
