+++ config/urls.py
@@ -20,7 +20,7 @@ handler500 = error500
 
 apps_patterns = [
     # path("", include("apps.users.urls")),
-    path("", include("apps.main.urls")),
+    path("", include("apps.core.urls")),
     path("", include("apps.afisha.urls")),
     # path("", include("apps.library.urls")),
     # path("", include("apps.articles.urls")),
